LayoutTests/js/dfg-to-string-valueOf-in-string-expected.txt

Tests that we do ToString conversions correctly when String.prototype.valueOf is fine but we define our own on the String object itself.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS "" + foo(argument) is "hello"
PASS successfullyParsed is true

TEST COMPLETE
